+++ idea/testData/refactoring/introduceParameter/coroutineLambdaWithSelection.kt
@@ -1,0 +1,7 @@
+fun foo(coroutine f: Int.() -> Continuation<Unit>) {
+
+}
+
+fun test() {
+    foo <selection>{}</selection>
+}