+++ idea/testData/refactoring/introduceProperty/coroutineLambdaNoSelection.kt
@@ -1,0 +1,8 @@
+// EXTRACTION_TARGET: property with getter
+fun foo(coroutine f: Int.() -> Continuation<Unit>) {
+
+}
+
+fun test() {
+    foo <caret>{}
+}