+++ idea/testData/smartSelection/coroutineLambda.kt
@@ -1,0 +1,11 @@
+fun foo(coroutine f: Int.() -> Continuation<Unit>) {
+
+}
+
+fun test() {
+    foo <caret>{}
+}
+
+/*
+foo {}
+*/