+++ ChangeLog.md
@@ -14,6 +14,7 @@
 - [`KT-14696`](https://youtrack.jetbrains.com/issue/KT-14696) Introduce Type Alias: Fix NPE during dialog repaint
 - [`KT-14685`](https://youtrack.jetbrains.com/issue/KT-14685) Introduce Type Alias: Replace type usages in constructor calls
 - [`KT-14861`](https://youtrack.jetbrains.com/issue/KT-14861) Introduce Type Alias: Support callable references/class literals
+- [`KT-14690`](https://youtrack.jetbrains.com/issue/KT-14690) Disable "Introduce..." refactorings on coroutine lambdas
 
 ## 1.1-M03 (EAP-3)